Update changelog date in README files
The most recent changes in the changelog add a new allowed key provided in the
data argument for the client.lists.members.update() and
client.lists.members.create_or_update() methods that does not require any
changes to those methods. The changelog section of the README files has been
updated to show that we are up to date with all changes up to that point.