Add basic support for oauth access_tokens
MailChimp now offers oauth2 support. The mechanism for using oauth
access tokens is a little different than the basic auth system used with
the old api_keys. This requires a call be made to the metadata_uri to
get the api endpoint with datacenter included. Then, requests simply
need to include an Authorization header of the format "Authorization:
OAuth <access_token>". Access tokens don't expire so there's no need to
deal with refreshing.