Update authentication to make user optional
This update is a breaking change from 2.0.x. The order of the MailChimpClient
initialization arguments for the basic auth has been reversed to accept the
api key first, and the username second, as the username argument is stated in
the documentation to accept any value. A default value of 'python-mailchimp'
has been provided in case the end user does not wish to use the username any
longer. An additional check was made to verify that the api key that was
entered is in the proper format and return an error if it is not as that is an
indication that the user did not reverse the order of their init as required.
The documentation has been updated to reflect this change. As this was a
breaking change, the name of the attribute used to authenticate with the api
token has been renamed from mc_secret to mc_api to provide a more accurate
name and will affect users that call these arguments explicitely.