Add missing error catching on MailChimpOAuth methods
Add a check for the token 'error' in the json output of the get_metadata()
method. This check will raise a RequestException with the value of the error
as the message. This also updates the get_base_url() method to reraise any
RequestException brought by using the get_metadata() method.